Initial client_participant split
